started implementation of tokenizer
